+++ app/src/main/java/com/oakkub/survey/data/repository/oauth/OAuthRepository.kt
@@ -1,6 +1,6 @@
 package com.oakkub.survey.data.repository.oauth
 
-import com.oakkub.survey.data.response.OAuthResponse
+import com.oakkub.survey.data.services.oauth.OAuthResponse
 import io.reactivex.Single
 
 /**
